Java_Advance

+++ intro/src/operator1.java
@@ -17,7 +17,7 @@ int intVal = 100;
 byte byteVal= (byte)intVal;
 byte max = 127;
 byte min = -128;
-byte sum= max + min;
+byte sum= (byte) (max + min);
 System.out.println("Sum= " + sum);
 }
 }
